refactor(handlers): prepare container module split
